refactor(chat): isolate session header controls
